worker/requirements.txt
requests>=2.32
notebooklm-py>=0.1.0
playwright>=1.47
python-pptx>=1.0
weasyprint>=62.0
markdown-it-py>=3.0
